+++ hpc-engine/cpca.c
@@ -117,10 +117,10 @@ int main( int argc , char ** argv )
     double rt;    
     
     // quantum friction
-    double qfalpha=0.0;
+    double qfswitch_it=0.0;
     
-    // current corrections
-    double cccoeff=0.0;
+    // variable used for particle number control
+    double pccoeff=0.0;
         
     /* start main */
     wt_b_t(); // tag init time
@@ -234,8 +234,13 @@ int main( int argc , char ** argv )
     // ====================================================================================
     // ======================== ALLOCATE GPU AND CPU BUFFERS ==============================
     // ====================================================================================
-    gpu_exec( host_malloc_pl((size_t)12*NXY*sizeof(double), (void **)&h_densities) );
-    gpu_exec(     gpu_malloc((size_t)12*NXY*sizeof(double), (void **)&d_densities) );
+    // quantum friction buffers
+    mpipackagesize=12;
+    if(md.qfalpha>0.0 || md.qfbeta>0.0 || md.qfgamma>0.0) mpipackagesize += 4;// quantum friction is expected to be used
+
+    gpu_exec( host_malloc_pl((size_t)mpipackagesize*NXY*sizeof(double), (void **)&h_densities) );
+    gpu_exec(     gpu_malloc((size_t)mpipackagesize*NXY*sizeof(double), (void **)&d_densities) );
+    double *d_qf_densities=d_densities + 12*NXY; // to simplify notation
     
     // potentials
     gpu_exec( host_malloc_pl((size_t)12*NXY*sizeof(double), (void **)&h_potentials) );
@@ -817,7 +822,7 @@ int main( int argc , char ** argv )
     modify_densities(it, densall, md.params, extra_data_size, extra_data, densall_d, d_extra_data);
 #endif
     // potentials
-//     if(md.inittype!=5) gpu_exec( compute_potentials(it, d_densities, d_potentials, cccoeff, md.nthreads) );
+//     if(md.inittype!=5) gpu_exec( compute_potentials(it, d_densities, d_potentials, 0.0, md.nthreads) );
     // energy
     gpu_exec( compute_energy(it, d_densities, d_potentials, d_workarea, md.nthreads) ); 
     
@@ -964,12 +969,11 @@ int main( int argc , char ** argv )
         {
             // ----------------------------- predictor -----------------------------------
             // compute value of quantum friction coefficient
-            qfalpha = 0.0; // NOTE: I assume there is no quantum friction during the first steps
-            cccoeff = h_smooth_step(t0+it*dt, md.ccstart/eF,  md.ccstop/eF,  md.ccswitch/eF, 1.0);
+            qfswitch_it = 0.0; // NOTE: I assume there is no quantum friction during the first steps
 
             // derivatives
 #ifdef FAST_CONST_EFFECTIVE_MASS_MODE
-            if(qfalpha>0.0) gradients_computed=1; else gradients_computed=0;
+            if(qfswitch_it>0.0) gradients_computed=1; else gradients_computed=0;
 #endif
             if(gradients_computed)
             {
@@ -997,7 +1001,7 @@ int main( int argc , char ** argv )
             modify_densities(0, densall, md.params, extra_data_size, extra_data, densall_d, d_extra_data);
 #endif
             // potentials - NOTE: it=0!
-            gpu_exec( compute_potentials(0, d_densities, d_potentials, cccoeff, md.nthreads) );
+            gpu_exec( compute_potentials(0, d_densities, d_potentials, 0.0, md.nthreads) );
 #ifndef FAST_CONST_EFFECTIVE_MASS_MODE
             // filtering of effective masses
             if(md.hkf_mode>=1)
@@ -1013,7 +1017,7 @@ int main( int argc , char ** argv )
             gpu_exec( memcopy_gpu2gpu(d_wf, d_fkm3, (size_t)2*nwfip*NXY*sizeof(double complex)) );
             gpu_exec( apply_hamiltonian(0, nwfip, d_fkm3, d_fkm1, /* NOTE - d_fkm1 as output buffer  */
                                     d_wf_d_dx, d_wf_d_dy, d_kkz, d_wf_laplace, d_alphawf_laplace,
-                                    d_densities, d_potentials, qfalpha, NULL, cccoeff, 
+                                    d_densities, d_potentials, qfswitch_it, NULL, pccoeff,
                                     md.nthreads) );
             // Make copy of qpe
             gpu_exec( memcopy_gpu2gpu(d_workarea, d_qpe, (size_t)nwfip*sizeof(double)) );
@@ -1045,7 +1049,7 @@ int main( int argc , char ** argv )
                 // H*psi
                 gpu_exec( apply_hamiltonian(0, nwfip, d_fkm2, d_fkm1,
                                         d_wf_d_dx, d_wf_d_dy, d_kkz, d_wf_laplace, d_alphawf_laplace,
-                                        d_densities, d_potentials, qfalpha, d_qpe, cccoeff, 
+                                        d_densities, d_potentials, qfswitch_it, d_qpe, pccoeff,
                                         md.nthreads) );
                 // Add contribution from Taylor expansion
                 gpu_exec( taylor_expansion_contribution(i_meas+1, 0.5*dt, nwfip, d_fkm1, d_fkm3, d_fkm2, md.nthreads) );
@@ -1058,12 +1062,11 @@ int main( int argc , char ** argv )
             
             // ----------------------------- corrector -----------------------------------
             // compute value of quantum friction coefficient
-            qfalpha = 0.0; // NOTE: I assume there is no quantum friction during the first steps
-            cccoeff = h_smooth_step(t0+it*dt, md.ccstart/eF,  md.ccstop/eF,  md.ccswitch/eF, 1.0);
+            qfswitch_it = 0.0; // NOTE: I assume there is no quantum friction during the first steps
 
             // derivatives
 #ifdef FAST_CONST_EFFECTIVE_MASS_MODE
-            if(qfalpha>0.0) gradients_computed=1; else gradients_computed=0;
+            if(qfswitch_it>0.0) gradients_computed=1; else gradients_computed=0;
 #endif
             if(gradients_computed)
             {
@@ -1091,7 +1094,7 @@ int main( int argc , char ** argv )
             modify_densities(0, densall, md.params, extra_data_size, extra_data, densall_d, d_extra_data);
 #endif
             // potentials - NOTE: it=0!
-            gpu_exec( compute_potentials(0, d_densities, d_potentials, cccoeff, md.nthreads) );
+            gpu_exec( compute_potentials(0, d_densities, d_potentials, 0.0, md.nthreads) );
             // NOTE - densities and potentials are computed for midpoint 
 #ifndef FAST_CONST_EFFECTIVE_MASS_MODE
             // filtering of effective masses
@@ -1124,7 +1127,7 @@ int main( int argc , char ** argv )
             // H*psi - first execution
             gpu_exec( apply_hamiltonian(0, nwfip, d_wf, d_fkm1, /* NOTE - d_fkm1 as output buffer  */
                                     d_wf_d_dx, d_wf_d_dy, d_kkz, d_wf_laplace, d_alphawf_laplace,
-                                    d_densities, d_potentials, qfalpha, NULL, cccoeff, 
+                                    d_densities, d_potentials, qfswitch_it, NULL, pccoeff,
                                     md.nthreads) );
             // Make copy of qpe
             gpu_exec( memcopy_gpu2gpu(d_workarea, d_qpe, (size_t)nwfip*sizeof(double)) );
@@ -1153,7 +1156,7 @@ int main( int argc , char ** argv )
                 // H*psi
                 gpu_exec( apply_hamiltonian(0, nwfip, d_fkm2, d_fkm1,
                                         d_wf_d_dx, d_wf_d_dy, d_kkz, d_wf_laplace, d_alphawf_laplace,
-                                        d_densities, d_potentials, qfalpha, d_qpe, cccoeff, 
+                                        d_densities, d_potentials, qfswitch_it, d_qpe, pccoeff,
                                         md.nthreads) );
                 // Add contribution from Taylor expansion
                 gpu_exec( taylor_expansion_contribution(i_meas+1, dt, nwfip, d_fkm1, d_wf, d_fkm2, md.nthreads) );
@@ -1210,7 +1213,7 @@ int main( int argc , char ** argv )
         modify_densities(it, densall, md.params, extra_data_size, extra_data, densall_d, d_extra_data);
 #endif
         // potentials
-        gpu_exec( compute_potentials(it, d_densities, d_potentials, cccoeff, md.nthreads) );
+        gpu_exec( compute_potentials(it, d_densities, d_potentials, 0.0, md.nthreads) );
         // energy
         gpu_exec( compute_energy(it, d_densities, d_potentials, d_workarea, md.nthreads) );
     
@@ -1257,8 +1260,7 @@ int main( int argc , char ** argv )
         {
             // ----------------------------- predictor -----------------------------------
             // compute value of quantum friction coefficient and current corrections coeff
-            qfalpha = md.qfalpha*h_smooth_step(t0+(it+1)*dt, md.qfstart/eF,  md.qfstop/eF,  md.qfswitch/eF, 1.0);
-            cccoeff = h_smooth_step(t0+(it+1)*dt, md.ccstart/eF,  md.ccstop/eF,  md.ccswitch/eF, 1.0);
+            qfswitch_it = quantum_friction_switch(t0+(it+1)*dt, eF);
 
             gpu_exec( abm_step1(nwfip, d_wf, d_fkm1, d_fkm2, d_fkm3, d_fkm4, d_fkm5, INTEGRATION_SCHEME, md.nthreads) );
 
@@ -1266,7 +1268,7 @@ int main( int argc , char ** argv )
             gpu_exec( normalize_wf(nwfip, d_wf, md.nthreads) );
             // derivatives
 #ifdef FAST_CONST_EFFECTIVE_MASS_MODE
-            if(qfalpha>0.0) gradients_computed=1; else gradients_computed=0;
+            if(qfswitch_it>0.0) gradients_computed=1; else gradients_computed=0;
 #endif
             if(gradients_computed)
             {
@@ -1279,7 +1281,12 @@ int main( int argc , char ** argv )
             // densities - local reduction
             gpu_exec( calculate_densities(nwfip, d_wf, d_wf_d_dx, d_wf_d_dy, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_densities, gradients_computed, md.nthreads) );
             // densities - global reduction
-            mpipackagesize = EXCHANGE_SIZE;
+            if(qfswitch_it>0.0) // quantum friction is on
+            {
+                gpu_exec( calculate_quantum_friction_densities(nwfip, d_wf, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_qf_densities, md.nthreads) );
+                mpipackagesize = 12+4;
+            }
+            else mpipackagesize = EXCHANGE_SIZE;
 #ifdef USE_GPU_AWARE_MPI
             MPI_Allreduce( MPI_IN_PLACE, d_densities, mpipackagesize*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
 #else
@@ -1295,7 +1302,7 @@ int main( int argc , char ** argv )
             modify_densities(it+1, densall, md.params, extra_data_size, extra_data, densall_d, d_extra_data);
 #endif
             // potentials
-            gpu_exec( compute_potentials(it+1, d_densities, d_potentials, cccoeff, md.nthreads) );
+            gpu_exec( compute_potentials(it+1, d_densities, d_potentials, 0.0, md.nthreads) );
 #ifndef FAST_CONST_EFFECTIVE_MASS_MODE
             // filtering of effective masses
             if(md.hkf_mode>=1)
@@ -1306,16 +1313,16 @@ int main( int argc , char ** argv )
             gpu_exec( multiply_wf_by_alpha(nwfip, d_wf, d_alphawf_laplace, d_potentials, md.nthreads) );
             gpu_exec( compute_laplace(2*nwfip, d_alphawf_laplace, d_alphawf_laplace, md.nthreads) );
 #endif
+            if(qfswitch_it>0.0) pccoeff=quantum_friction_pccoeff(NXY,densall.rho_a,densall.rho_b,DX*DY*LZ); // see pca_utils.c
             // H*psi
             gpu_exec( apply_hamiltonian(it+1, nwfip, d_wf, d_wf_laplace, /* NOTE - d_wf_laplace as output buffer  */
                                     d_wf_d_dx, d_wf_d_dy, d_kkz, d_wf_laplace, d_alphawf_laplace,
-                                    d_densities, d_potentials, qfalpha, NULL, cccoeff, 
+                                    d_densities, d_potentials, qfswitch_it, NULL, pccoeff,
                                     md.nthreads) ); 
             
             // ----------------------------- corrector -----------------------------------
             // compute value of quantum friction coefficient  and current corrections coeff
-            qfalpha = md.qfalpha*h_smooth_step(t0+(it+1)*dt, md.qfstart/eF,  md.qfstop/eF,  md.qfswitch/eF, 1.0);
-            cccoeff = h_smooth_step(t0+(it+1)*dt, md.ccstart/eF,  md.ccstop/eF,  md.ccswitch/eF, 1.0);
+            qfswitch_it = quantum_friction_switch(t0+(it+1)*dt, eF);
 
 #if INTEGRATION_SCHEME==AB3AM4
             gpu_exec( abm_step4(nwfip, d_wf_laplace, /* NOTE - d_wf_laplace as itermiediate buffer  */
@@ -1332,7 +1339,7 @@ int main( int argc , char ** argv )
             gpu_exec( normalize_wf(nwfip, d_wf, md.nthreads) );
             // derivatives
 #ifdef FAST_CONST_EFFECTIVE_MASS_MODE
-            if(qfalpha>0.0 || i_step==md.timesteps-1) gradients_computed=1; else gradients_computed=0;
+            if(qfswitch_it>0.0 || i_step==md.timesteps-1) gradients_computed=1; else gradients_computed=0;
 #endif
             if(gradients_computed)
             {
@@ -1345,7 +1352,13 @@ int main( int argc , char ** argv )
             // densities - local reduction
             gpu_exec( calculate_densities(nwfip, d_wf, d_wf_d_dx, d_wf_d_dy, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_densities, gradients_computed, md.nthreads) );
             // densities - global reduction
-            if(i_step==md.timesteps-1) mpipackagesize = 12; else mpipackagesize = EXCHANGE_SIZE;
+            if(qfswitch_it>0.0) // quantum friction is on
+            {
+                gpu_exec( calculate_quantum_friction_densities(nwfip, d_wf, d_wf_laplace, d_kkz, d_fbetaEn, NULL, d_qf_densities, md.nthreads) );
+                mpipackagesize = 12+4;
+            }
+            else if(i_step==md.timesteps-1) mpipackagesize = 12;
+            else mpipackagesize = EXCHANGE_SIZE;
 #ifdef USE_GPU_AWARE_MPI
             MPI_Allreduce( MPI_IN_PLACE, d_densities, mpipackagesize*NXY, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
 #else
@@ -1373,7 +1386,7 @@ int main( int argc , char ** argv )
                 file_operation( write_measurments_subset(&wdmd, MPI_COMM_WORLD, "td", it, densall_subset) );
             }
             // potentials
-            gpu_exec( compute_potentials(it+1, d_densities, d_potentials, cccoeff, md.nthreads) );
+            gpu_exec( compute_potentials(it+1, d_densities, d_potentials, 0.0, md.nthreads) );
 #ifndef FAST_CONST_EFFECTIVE_MASS_MODE
             // filtering of effective masses
             if(md.hkf_mode>=1)
@@ -1405,10 +1418,11 @@ int main( int argc , char ** argv )
             d_fkm2=d_fkm1;
             d_fkm1=d_tmp_ptr;
 #endif
+            if(qfswitch_it>0.0) pccoeff=quantum_friction_pccoeff(NXY,densall.rho_a,densall.rho_b,DX*DY*LZ); // see pca_utils.c
             // H*psi
             gpu_exec( apply_hamiltonian(it+1, nwfip, d_wf, d_fkm1,
                                     d_wf_d_dx, d_wf_d_dy, d_kkz, d_wf_laplace, d_alphawf_laplace, 
-                                    d_densities, d_potentials, qfalpha, NULL, cccoeff, 
+                                    d_densities, d_potentials, qfswitch_it, NULL, pccoeff,
                                     md.nthreads) );            
             
             it++; // update global time counter