+++ hpc-engine/cpca_kernels.cu
@@ -233,7 +233,6 @@ __global__ void kernel_apply_hamiltonian(int it, wslda_potential h_potentials,
                                          double *j_corr_a_x, double *j_corr_a_y, double *j_corr_a_z, double *j_corr_b_x, double *j_corr_b_y, double *j_corr_b_z,
                                          size_t n, Complex *wf_in, Complex *wf_out,
                                          Complex *wf_d_dx, Complex *wf_d_dy, double *d_kkz, Complex *wf_laplace, Complex *alphawf_laplace,
-                                         double cccoeff,
                                          double *vx_a, double *vy_a, double *divv_a, double *vx_b, double *vy_b, double *divv_b
                                         )
 {
@@ -491,20 +490,56 @@ __global__ void kernel_apply_hamiltonian_bdg(int it,
 }
 
 __global__ void kernel_add_quantum_friction(double *rho_a, double *rho_b,
-                                            double *djax_dx, double *djay_dy, double *djaz_dz, double *djbx_dx, double *djby_dy, double *djbz_dz,
-                                            double *V_a, double *V_b, double qfalpha)
+                                             double* d_qf_density_for_Ua, double* d_qf_density_for_Ub, Complex*d_qf_density_for_D,
+                                             double *V_a, double *V_b, Complex *delta,
+                                            double qfalpha, double qfbeta, double qfgamma)
 {
     size_t ixyz= threadIdx.x + blockIdx.x * blockDim.x; // compute for this point
 
-    int ix, iy;
-    double coeff=0.0, r;
+    double PhDel,PhDen;
+    double _V_a, _V_b;
+    Complex _delta;
     if(ixyz<NXY)
     {
         // see Eq.(3) in paper https://arxiv.org/abs/1305.6891
-        V_a[ixyz]-=qfalpha*(djax_dx[ixyz]+djay_dy[ixyz])/dc_nF; // here I divide be reference density, to avoid problems of division by zero
-        V_b[ixyz]-=qfalpha*(djbx_dx[ixyz]+djby_dy[ixyz])/dc_nF; // here I divide be reference density, to avoid problems of division by zero
+        //V_a[ixyz]-=qfalpha*(djax_dx[ixyz]+djay_dy[ixyz])/dc_nF; // here I divide be reference density, to avoid problems of division by zero
+        //V_b[ixyz]-=qfalpha*(djbx_dx[ixyz]+djby_dy[ixyz])/dc_nF; // here I divide be reference density, to avoid problems of division by zero
+
+                // // see Eq.(3) in paper https://arxiv.org/abs/1305.6891 //---------EA VERSION
+        PhDel         = thrust::arg(delta[ixyz]);
+        PhDen         = thrust::arg(d_qf_density_for_D[ixyz]);
+
+        _V_a = V_a[ixyz]; // keep the original value
+        V_a[ixyz]   -=d_qf_density_for_Ua[ixyz]*qfalpha/dc_nF; // here I divide be reference density, to avoid problems of division by zero
+        _V_b = V_b[ixyz]; // keep the original value
+        V_b[ixyz]   -=d_qf_density_for_Ub[ixyz]*qfalpha/dc_nF; // here I divide be reference density, to avoid problems of division by zero
+        _delta = delta[ixyz]; // keep the original value
+        delta[ixyz] +=qfbeta*thrust::abs(delta[ixyz])*Complex(cos(PhDel),sin(PhDel))*sin(PhDel - PhDen); //UWAGA NOT FINAL FORM as its missing the Non particle conserving term - \gamma *[N(t)-N_req] \Delta
+        // particle control part
+        delta[ixyz] +=Complex(0.0,qfgamma)*_delta;
+
+        // Save original values - it will be used by kernel_remove_quantum_friction(...)
+        d_qf_density_for_Ua[ixyz]=_V_a;
+        d_qf_density_for_Ub[ixyz]=_V_b;
+        d_qf_density_for_D[ixyz]=_delta;
     }
 }
+
+
+__global__ void kernel_remove_quantum_friction(double* d_qf_density_for_Ua, double* d_qf_density_for_Ub, Complex*d_qf_density_for_D,
+                                             double *V_a, double *V_b, Complex *delta)
+{
+    size_t ixyz= threadIdx.x + blockIdx.x * blockDim.x; // compute for this point
+    if(ixyz<NXY)
+    {
+        // restore values without quantum friction
+        V_a[ixyz]=d_qf_density_for_Ua[ixyz];
+        V_b[ixyz]=d_qf_density_for_Ub[ixyz];
+        delta[ixyz]=d_qf_density_for_D[ixyz];
+    }
+}
+
+
 
 __global__ void kernel_compute_qpe(Complex *wf1_u, Complex *wf2_u, Complex *wf1_v, Complex *wf2_v, double *re)
 {
@@ -630,18 +665,17 @@ __global__ void kernel_get_vector_vext(int it, int spin, double *vx, double *vy)
  *                     where: V_a, V_b - double arrays of size NXY
  *                            delta - double complex array of size NXY
  *                     In total size of d_potentials is 4*NXY
- * @param qfalpha coefficient for quantum friction term, beta coefficient in Eq.(3) in paper https://arxiv.org/abs/1305.6891,
- *                if qfalpha=0.0 then quantum friction is NOT active
+ * @param qfswitch switch coefficient for quantum friction term, if qfswitch=0.0 then quantum friction is NOT active
  * @param useqpe array of size [n]
  *               if NULL then quasiparticle energies will be computed from wf_in,
  *               otherwise given array will be used,
- * @param cccoeff the current corrections coefficient
+ * @param pccoeff the particle control coefficient: N(t) - N_req
  * @param nthreads number of threads per block
  * @return 0 - OK, otherwise ERROR
  * */
 extern "C" int apply_hamiltonian(int it, int n, cufftDoubleComplex *wf_in, cufftDoubleComplex *wf_out,
                             cufftDoubleComplex *wf_d_dx, cufftDoubleComplex *wf_d_dy, double *d_kkz, cufftDoubleComplex *wf_laplace, cufftDoubleComplex *alphawf_laplace,
-                            double *d_densities, double *d_potentials, double qfalpha, double *useqpe, double cccoeff,
+                            double *d_densities, double *d_potentials, double qfswitch, double *useqpe, double pccoeff,
                             int nthreads)
 {
     // number of blocks
@@ -679,9 +713,21 @@ extern "C" int apply_hamiltonian(int it, int n, cufftDoubleComplex *wf_in, cufft
     double * laplace_alpha_b  = grad_alpha_a + NXY*13;
 
     // Step 1: if quantum friction is active, update mean-field potentials
-    if(qfalpha>0.0)
+    if(qfswitch>0.0)
     {
-//         printf("QUANTUM FRICTION ACTIVE! qfalpha=%f\n", qfalpha);
+        double qfalpha = md.qfalpha*qfswitch;
+        double qfbeta = md.qfbeta*qfswitch;
+        double qfgamma = md.qfgamma*qfswitch*pccoeff;
+
+        double * d_qf_density_for_Ua = (double *) (d_densities+12*NXY);  // density for diagonal part (U) of quantum friction force
+        double * d_qf_density_for_Ub = (double *) (d_qf_density_for_Ua+1*NXY);
+        Complex *d_qf_density_for_D = (Complex *) (d_qf_density_for_Ua+2*NXY); // density for off-diagonal part (Delta) of quantum friction force
+
+        // TODO: EA: Update this section
+        // TODO: For now I leave the old method, but you should replace it with computation via second derivatives
+        // TODO: Here you need to update mean-field potentials (V_a,V_b) and pairing potential (Delta)
+//**************************************OLD METHOD OF USING THE GRADIENT OF THE CURRENTS**************************************  //---------Gw VERSION
+/* 
         // compute nabla*j, use grad_j_corr_a and grad_j_corr_b as temporary buffers
         ierr=compute_derivative_real_vector_f(j_a_x, j_a_y, NULL, grad_j_corr_a, grad_j_corr_a+NXY, NULL, nthreads);
         if(ierr!=0) return ierr;
@@ -692,6 +738,16 @@ extern "C" int apply_hamiltonian(int it, int n, cufftDoubleComplex *wf_in, cufft
         kernel_add_quantum_friction<<<nblocks, nthreads>>>(rho_a, rho_b,
                                                     grad_j_corr_a, grad_j_corr_a+NXY, NULL, grad_j_corr_b, grad_j_corr_b+NXY, NULL,
                                                     V_a, V_b, qfalpha);
+ */
+       //******************************************NEW METHOD OF LAPLACIAN OF WF******************************************   //---------EA VERSION
+   
+            // update mean field potential by friction terms
+
+           kernel_add_quantum_friction<<<nblocks, nthreads>>>(rho_a, rho_b,
+                                                    d_qf_density_for_Ua, d_qf_density_for_Ub, d_qf_density_for_D,
+                                                    V_a, V_b, delta, qfalpha, qfbeta, qfgamma);
+
+
     }
 
     // filtering of mean-fields
@@ -765,7 +821,6 @@ extern "C" int apply_hamiltonian(int it, int n, cufftDoubleComplex *wf_in, cufft
                                             grad_j_corr_a, grad_j_corr_a+NXY, NULL, grad_j_corr_b, grad_j_corr_b+NXY, NULL,
                                             n, (Complex *)wf_in, (Complex *)wf_out,
                                             (Complex *)wf_d_dx, (Complex *)wf_d_dy, d_kkz, (Complex *)wf_laplace, (Complex *)alphawf_laplace,
-                                            cccoeff,
                                             vecvext_a, vecvext_a+NXY, divvext_a, vecvext_b, vecvext_b+NXY, divvext_b
                                                    );
 #endif
@@ -799,6 +854,20 @@ extern "C" int apply_hamiltonian(int it, int n, cufftDoubleComplex *wf_in, cufft
     nblocks = (int)ceil((float)NXY/nthreads);
     kernel_subtruct_qpe<<<nblocks, nthreads>>>(n, (Complex *)wf_in, (Complex *)wf_out, gpe);  
  
+
+    // if quantum friction was active, remove contributions to mean-fields
+    if(qfswitch>0.0)
+    {
+        double * d_qf_density_for_Ua = (double *) (d_densities+12*NXY);  // density for diagonal part (U) of quantum friction force
+        double * d_qf_density_for_Ub = (double *) (d_qf_density_for_Ua+1*NXY);
+        Complex *d_qf_density_for_D = (Complex *) (d_qf_density_for_Ua+2*NXY); // density for off-diagonal part (Delta) of quantum
+
+        nblocks = (int)ceil((float)NXY/nthreads);
+        //kernel_remove_quantum_friction<<<nblocks, nthreads>>>(d_qf_density_for_Ua, d_qf_density_for_Ub, d_qf_density_for_D,
+        //                                                      V_a, V_b, delta);
+    }
+
+
     return 0;
 }