+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 2025.xx.xx
+* Implementation of improved quantum friction functionality.
+* Added new tests.
+* Minor bug fixes and improvements.
+
 # 2024.08.11
 * Improved stability of td codes (filtering of noise from mean-fields).
 * Update for new API version of ELPA library.